# docs/FHIRAPI-swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "FHIR API",
    "description": "FHIR R4 and R5 view of the EHR. Sections show who calls each endpoint: payers during CRD, payer DTR apps, and the Payer Gateway (/internal).",
    "version": "v1"
  },
  "paths": {
...
    "/fhir/r4/QuestionnaireResponse": {
      "post": {
        "tags": [
          "2. DTR forms (QuestionnaireResponse) - saved by the Payer Gateway or a payer DTR app"
        ],
        "summary": "Creates a new form. Any id in the body is replaced by a server id. Returns 201 + Location header.",
        "description": "Must be filled in the body: resourceType = \"QuestionnaireResponse\" and subject.reference = \"Patient/{patient id}\".\nThe order and the coverage the form is for are read from the DTR \"qr-context\" extensions, when present.",
        "requestBody": {
          "description": "Required. A FHIR QuestionnaireResponse (subject = Patient/{id} is required).",
          "content": {
            "application/fhir+json": {
              "schema": { },
              "example": {
                "resourceType": "QuestionnaireResponse",
                "questionnaire": "http://example.org/fhir/Questionnaire/dialysis-incenter-hd",
                "status": "in-progress",
                "subject": {
                  "reference": "Patient/pat-1"
                },
                "authored": "2026-10-07T18:36:05Z",
                "author": {
                  "reference": "Practitioner/prac-1"
                },
                "extension": [
                  {
                    "url": "http://hl7.org/fhir/us/davinci-dtr/StructureDefinition/qr-context",
                    "valueReference": {
                      "reference": "ServiceRequest/ord-1-hd"
                    }
                  },
                  {
                    "url": "http://hl7.org/fhir/us/davinci-dtr/StructureDefinition/qr-context",
                    "valueReference": {
                      "reference": "Coverage/cov-1"
                    }
                  }
                ],
                "item": [
                  {
                    "linkId": "egfr",
                    "text": "Most recent eGFR (mL/min/1.73 m2)",
                    "answer": [
                      {
                        "valueDecimal": 9
                      }
                    ]
                  }
                ]
              }
            },
            "application/json": {
              "schema": { },
              "example": {
                "resourceType": "QuestionnaireResponse",
                "questionnaire": "http://example.org/fhir/Questionnaire/dialysis-incenter-hd",
                "status": "in-progress",
                "subject": {
                  "reference": "Patient/pat-1"
                },
                "authored": "2026-10-07T18:36:05Z",
                "author": {
                  "reference": "Practitioner/prac-1"
                },
                "extension": [
                  {
                    "url": "http://hl7.org/fhir/us/davinci-dtr/StructureDefinition/qr-context",
                    "valueReference": {
                      "reference": "ServiceRequest/ord-1-hd"
                    }
                  },
                  {
                    "url": "http://hl7.org/fhir/us/davinci-dtr/StructureDefinition/qr-context",
                    "valueReference": {
                      "reference": "Coverage/cov-1"
                    }
                  }
                ],
                "item": [
                  {
                    "linkId": "egfr",
                    "text": "Most recent eGFR (mL/min/1.73 m2)",
                    "answer": [
                      {
                        "valueDecimal": 9
                      }
                    ]
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Saved. Body: the form as stored, with its new id; the Location header has its URL."
          },
          "400": {
            "description": "The body is not a QuestionnaireResponse, or subject does not point to a patient. Body: a FHIR OperationOutcome, e.g. { \"resourceType\": \"OperationOutcome\", \"issue\": [ { \"severity\": \"error\", \"code\": \"required\", \"diagnostics\": \"QuestionnaireResponse.subject must reference a Patient, e.g. 'Patient/pat-1'.\" } ] }."
          },
          "502": {
            "description": "The EHR data API could not be reached or answered with an error. Body: a FHIR OperationOutcome."
          }
        }
      },
      "get": {
        "tags": [
          "2. DTR forms (QuestionnaireResponse) - saved by the Payer Gateway or a payer DTR app"
        ],
        "summary": "Returns all forms of a patient as a searchset Bundle (newest first).",
        "parameters": [
          {
            "name": "patient",
            "in": "query",
            "description": "Required. Patient id or \"Patient/id\", e.g. pat-1.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A FHIR Bundle of type \"searchset\" with the forms of the patient (total 0 and no entry when there are none)."
          },
          "400": {
            "description": "The patient parameter is missing. Body: a FHIR OperationOutcome, e.g. { \"resourceType\": \"OperationOutcome\", \"issue\": [ { \"severity\": \"error\", \"code\": \"required\", \"diagnostics\": \"QuestionnaireResponse search requires the patient parameter.\" } ] }."
          },
          "502": {
            "description": "The EHR data API could not be reached or answered with an error. Body: a FHIR OperationOutcome."
          }
        }
      }
    },
    "/fhir/r5/QuestionnaireResponse": {
      "post": {
        "tags": [
          "2. DTR forms (QuestionnaireResponse) - saved by the Payer Gateway or a payer DTR app"
        ],
        "summary": "Creates a new form. Any id in the body is replaced by a server id. Returns 201 + Location header.",
        "description": "Must be filled in the body: resourceType = \"QuestionnaireResponse\" and subject.reference = \"Patient/{patient id}\".\nThe order and the coverage the form is for are read from the DTR \"qr-context\" extensions, when present.",
        "requestBody": {
          "description": "Required. A FHIR QuestionnaireResponse (subject = Patient/{id} is required).",
          "content": {
            "application/fhir+json": {
              "schema": { },
              "example": {
                "resourceType": "QuestionnaireResponse",
                "questionnaire": "http://example.org/fhir/Questionnaire/dialysis-incenter-hd",
                "status": "in-progress",
                "subject": {
                  "reference": "Patient/pat-1"
                },
                "authored": "2026-10-07T18:36:05Z",
                "author": {
                  "reference": "Practitioner/prac-1"
                },
                "extension": [
                  {
                    "url": "http://hl7.org/fhir/us/davinci-dtr/StructureDefinition/qr-context",
                    "valueReference": {
                      "reference": "ServiceRequest/ord-1-hd"
                    }
                  },
                  {
                    "url": "http://hl7.org/fhir/us/davinci-dtr/StructureDefinition/qr-context",
                    "valueReference": {
                      "reference": "Coverage/cov-1"
                    }
                  }
                ],
                "item": [
                  {
                    "linkId": "egfr",
                    "text": "Most recent eGFR (mL/min/1.73 m2)",
                    "answer": [
                      {
                        "valueDecimal": 9
                      }
                    ]
                  }
                ]
              }
            },
            "application/json": {
              "schema": { },
              "example": {
                "resourceType": "QuestionnaireResponse",
                "questionnaire": "http://example.org/fhir/Questionnaire/dialysis-incenter-hd",
                "status": "in-progress",
                "subject": {
                  "reference": "Patient/pat-1"
                },
                "authored": "2026-10-07T18:36:05Z",
                "author": {
                  "reference": "Practitioner/prac-1"
                },
                "extension": [
                  {
                    "url": "http://hl7.org/fhir/us/davinci-dtr/StructureDefinition/qr-context",
                    "valueReference": {
                      "reference": "ServiceRequest/ord-1-hd"
                    }
                  },
                  {
                    "url": "http://hl7.org/fhir/us/davinci-dtr/StructureDefinition/qr-context",
                    "valueReference": {
                      "reference": "Coverage/cov-1"
                    }
                  }
                ],
                "item": [
                  {
                    "linkId": "egfr",
                    "text": "Most recent eGFR (mL/min/1.73 m2)",
                    "answer": [
                      {
                        "valueDecimal": 9
                      }
                    ]
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Saved. Body: the form as stored, with its new id; the Location header has its URL."
          },
          "400": {
            "description": "The body is not a QuestionnaireResponse, or subject does not point to a patient. Body: a FHIR OperationOutcome, e.g. { \"resourceType\": \"OperationOutcome\", \"issue\": [ { \"severity\": \"error\", \"code\": \"required\", \"diagnostics\": \"QuestionnaireResponse.subject must reference a Patient, e.g. 'Patient/pat-1'.\" } ] }."
          },
          "502": {
            "description": "The EHR data API could not be reached or answered with an error. Body: a FHIR OperationOutcome."
          }
        }
      },
      "get": {
        "tags": [
          "2. DTR forms (QuestionnaireResponse) - saved by the Payer Gateway or a payer DTR app"
        ],
        "summary": "Returns all forms of a patient as a searchset Bundle (newest first).",
        "parameters": [
          {
            "name": "patient",
            "in": "query",
            "description": "Required. Patient id or \"Patient/id\", e.g. pat-1.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A FHIR Bundle of type \"searchset\" with the forms of the patient (total 0 and no entry when there are none)."
          },
          "400": {
            "description": "The patient parameter is missing. Body: a FHIR OperationOutcome, e.g. { \"resourceType\": \"OperationOutcome\", \"issue\": [ { \"severity\": \"error\", \"code\": \"required\", \"diagnostics\": \"QuestionnaireResponse search requires the patient parameter.\" } ] }."
          },
          "502": {
            "description": "The EHR data API could not be reached or answered with an error. Body: a FHIR OperationOutcome."
          }
        }
      }
    },
    "/fhir/r4/QuestionnaireResponse/{id}": {
      "put": {
        "tags": [
          "2. DTR forms (QuestionnaireResponse) - saved by the Payer Gateway or a payer DTR app"
        ],
        "summary": "Creates (201) or replaces (200) the form with this id. The body id, if present, must equal the URL id.",
        "description": "Must be filled in the body: resourceType = \"QuestionnaireResponse\" and subject.reference = \"Patient/{patient id}\".\nSend the whole form each time: it replaces the stored one.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Required. Form id chosen by the caller, e.g. \"qr-frm-235b7ad69e85\".",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "qr-frm-235b7ad69e85"
          }
        ],
        "requestBody": {
          "description": "Required. A FHIR QuestionnaireResponse.",
          "content": {
            "application/fhir+json": {
              "schema": { },
              "example": {
                "resourceType": "QuestionnaireResponse",
                "questionnaire": "http://example.org/fhir/Questionnaire/dialysis-incenter-hd",
                "status": "in-progress",
                "subject": {
                  "reference": "Patient/pat-1"
                },
                "authored": "2026-10-07T18:36:05Z",
                "author": {
                  "reference": "Practitioner/prac-1"
                },
                "extension": [
                  {
                    "url": "http://hl7.org/fhir/us/davinci-dtr/StructureDefinition/qr-context",
                    "valueReference": {
                      "reference": "ServiceRequest/ord-1-hd"
                    }
                  },
                  {
                    "url": "http://hl7.org/fhir/us/davinci-dtr/StructureDefinition/qr-context",
                    "valueReference": {
                      "reference": "Coverage/cov-1"
                    }
                  }
                ],
                "item": [
                  {
                    "linkId": "egfr",
                    "text": "Most recent eGFR (mL/min/1.73 m2)",
                    "answer": [
                      {
                        "valueDecimal": 9
                      }
                    ]
                  }
                ]
              }
            },
            "application/json": {
              "schema": { },
              "example": {
                "resourceType": "QuestionnaireResponse",
                "questionnaire": "http://example.org/fhir/Questionnaire/dialysis-incenter-hd",
                "status": "in-progress",
                "subject": {
                  "reference": "Patient/pat-1"
                },
                "authored": "2026-10-07T18:36:05Z",
                "author": {
                  "reference": "Practitioner/prac-1"
                },
                "extension": [
                  {
                    "url": "http://hl7.org/fhir/us/davinci-dtr/StructureDefinition/qr-context",
                    "valueReference": {
                      "reference": "ServiceRequest/ord-1-hd"
                    }
                  },
                  {
                    "url": "http://hl7.org/fhir/us/davinci-dtr/StructureDefinition/qr-context",
                    "valueReference": {
                      "reference": "Coverage/cov-1"
                    }
                  }
                ],
                "item": [
                  {
                    "linkId": "egfr",
                    "text": "Most recent eGFR (mL/min/1.73 m2)",
                    "answer": [
                      {
                        "valueDecimal": 9
                      }
                    ]
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Replaced. Body: the form as stored."
          },
          "201": {
            "description": "Created (the id was new). Body: the form as stored."
          },
          "400": {
            "description": "The body is not a QuestionnaireResponse, or subject does not point to a patient, or the id in the body is not the id in the URL. Body: a FHIR OperationOutcome, e.g. { \"resourceType\": \"OperationOutcome\", \"issue\": [ { \"severity\": \"error\", \"code\": \"required\", \"diagnostics\": \"QuestionnaireResponse.subject must reference a Patient, e.g. 'Patient/pat-1'.\" } ] }."
          },
          "502": {
            "description": "The EHR data API could not be reached or answered with an error. Body: a FHIR OperationOutcome."
          }
        }
      },
      "get": {
        "tags": [
          "2. DTR forms (QuestionnaireResponse) - saved by the Payer Gateway or a payer DTR app"
        ],
        "summary": "Returns one stored form exactly as it was saved.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Required. Form id.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "qr-frm-235b7ad69e85"
          }
        ],
        "responses": {
          "200": {
            "description": "The FHIR QuestionnaireResponse (content type application/fhir+json)."
          },
          "404": {
            "description": "No form with this id. Body: a FHIR OperationOutcome, e.g. { \"resourceType\": \"OperationOutcome\", \"issue\": [ { \"severity\": \"error\", \"code\": \"not-found\", \"diagnostics\": \"QuestionnaireResponse/qr-9 was not found.\" } ] }."
          },
          "502": {
            "description": "The EHR data API could not be reached or answered with an error. Body: a FHIR OperationOutcome."
          }
        }
      }
    },
...
  },
...
}